fix component naming, code style
