+++ src/Header/containers/ChannelName.tsx
@@ -1,13 +1,15 @@
-import {connect} from "react-redux";
-import {Header} from "../components/Header";
+import {connect} from 'react-redux';
+import {IState} from '../../common/IState';
+import {IChannel} from '../../Channels/models/IChannel';
+import {ChannelName, ICHannelNameStateProps} from '../components/ChannelName';
 
-// const mapStateToProps = (state: IState): ICHannelNameStateProps => {
-//     const channel = state.channelList.channels.find(
-//         (i: IChannel) => i.id === state.channelList.selectedChannelId
-//     );
-//     return {
-//         channelName: channel ? channel.name : null
-//     };
-// };
+const mapStateToProps = (state: IState): ICHannelNameStateProps => {
+    const channel = state.channelList.channels.find(
+        (i: IChannel) => i.id === state.channelList.selectedChannelId
+    );
+    return {
+        channelName: channel ? channel.name : null
+    };
+};
 
-export const HeaderContainer = connect()(Header);
+export const ChannelNameContainer = connect<ICHannelNameStateProps>(mapStateToProps)(ChannelName);
